Towards 2.0: move dependencies to submodules
updated libeba, libecheck, libehstr

The new version of E(asy)Check contains the E(asy)Embed library,
which removes the need for almost all platform detection code in
libehbigint. Also no longer have the source files for the deps
in the src/ dir (but rather in the submodules).

Collapsed the remaining headers and src files into a single pair of
src/ehbigint.h and src/ehbigint.c files.

Added ehbigint_arduino_tests ... skipping GMP tests naturally as
well as the "is_probably_prime" as that requires more dynamic memory
than is available on many arduinos. Left the "test_sqrt" test in,
even though it is pretty slow.

Have not yet adjusted the API to be more consistent with the error
as a passed-in parameter rather than a return value.